backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.1
python-dotenv==1.0.1
faiss-cpu==1.8.0
sentence-transformers==3.0.1
markdown==3.6
beautifulsoup4==4.12.3
requests==2.32.3
tqdm==4.66.5
sqlite-utils==3.36
cerebras-cloud-sdk==1.7.0
httpx==0.27.0
httpcore==1.0.4